refactor: set gridCategories as view
