+++ src/views/GridCategories.jsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import PropTypes from 'prop-types';
-import Card from './Card';
+import Card from '../components/Card';
 import styles from '../styles/GridCategories.module.scss';
 
 const GridCategories = ({data}) => {
